Fix disappearing CKEditor toolbar issue.
The toolbar would not appear on the Post tab if the Story tab
was visited first (thus getting to create its CKEditor first).

Fix was to create unique dom ids for the spaces. This is a bit of
the "quick" fix. A better fix would be to only create one post
editor and to reuse it in multiple contexts.